+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 ### Added
 
-Added Gemini single line autocomplete
+- Added Gemini single line autocomplete
 
 
+## [1.0.4] - 2024-07-21
+
+### Added
+
+- Added multiline context gemini with short cut `alt + g` or `> Gemini Assistant: Activate Gemini Assistant`
+- Assistant automatically on on all supported files.